+++ src/main/java/com/parse/gwt/client/types/Relation.java
@@ -1,0 +1,8 @@
+package com.parse.gwt.client.types;
+
+/**
+ * Created by Kerby on 2/23/2017.
+ */
+public class Relation {
+    public String className;
+}